+++ server/apps/notification-service/src/emails/email.module.ts
