+++ server/apps/notification-service/src/emails/email.service.ts
