+++ server/apps/notification-service/src/main.ts
@@ -3,6 +3,6 @@ import { NotificationServiceModule } from './notification-service.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(NotificationServiceModule);
-  await app.listen(process.env.port ?? 3400);
+  await app.listen(process.env.port ?? 3005);
 }
 bootstrap();
