+++ server/apps/notification-service/src/notification-service.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { NotificationService } from './notification-service.service';
 import { EventHandlers } from './events/event.handlers';
-import { EmailModule } from './email/email.module';
+import { EmailModule } from './emails/email.module';
 import { RabbitMQModule, UserServiceClient, UserServiceClientModule } from 'libs/shared-lib/src';
 import { HttpModule } from '@nestjs/axios';
